toml: moved functions for toml-loader to config/load.go.
