+++ config/toml.go
@@ -16,14 +16,8 @@ package config
 
 import (
 	"errors"
-	"fmt"
-	"github.com/BurntSushi/toml"
-	log "github.com/Sirupsen/logrus"
-	"io/ioutil"
 	"reflect"
 	"regexp"
-	"strconv"
-	"strings"
 	"unicode/utf8"
 )
 
@@ -90,149 +84,6 @@ func String(str NilOrString) string {
 	return stringValue
 }
 
-func getGatewayName(gatewaySectionMap SectionMap) string {
-	for name, value := range gatewaySectionMap {
-		if name == "name" {
-			gatewayName := value.(string)
-			return gatewayName
-		}
-	}
-	return ""
-}
-
-func addGatewaySection(configSections []ConfigSection, gatewaySectionMap SectionMap) []ConfigSection {
-	valueMap := make(ValueMap)
-	for name, value := range gatewaySectionMap {
-
-		switch value.(type) {
-		case int64:
-			valueMap[name] = strconv.FormatInt(value.(int64), 10)
-		case bool:
-			valueMap[name] = strconv.FormatBool(value.(bool))
-		default:
-			valueMap[name] = value.(string)
-		}
-
-	}
-
-	if len(valueMap) > 0 {
-		rt := ConfigSection{
-			Title:  "gateway",
-			Type:   "gateway",
-			Values: valueMap,
-		}
-		configSections = append(configSections, rt)
-	}
-
-	return configSections
-}
-
-func addStatusSections(configSections []ConfigSection, statusSectionMap SectionMap) []ConfigSection {
-	valueMap := make(ValueMap)
-	for name, value := range statusSectionMap {
-
-		switch value.(type) {
-		case int64:
-			valueMap[name] = strconv.FormatInt(value.(int64), 10)
-		case bool:
-			valueMap[name] = strconv.FormatBool(value.(bool))
-		case []map[string]interface{}:
-			// do nothing
-		default:
-			valueMap[name] = value.(string)
-		}
-
-	}
-	if len(valueMap) > 0 {
-		rt := ConfigSection{
-			Title:  "status",
-			Type:   "status",
-			Values: valueMap,
-		}
-		configSections = append(configSections, rt)
-	}
-
-	for name, value := range statusSectionMap {
-
-		valueMap := make(ValueMap)
-		switch value.(type) {
-		case []map[string]interface{}:
-			{
-				m := value.([]map[string]interface{})
-				for _, v := range m {
-					for k, vv := range v {
-						switch vv.(type) {
-						case int64:
-							valueMap[k] = strconv.FormatInt(vv.(int64), 10)
-						case bool:
-							valueMap[k] = strconv.FormatBool(vv.(bool))
-						default:
-							valueMap[k] = vv.(string)
-						}
-					}
-				}
-			}
-		}
-
-		if len(valueMap) > 0 {
-			rt := ConfigSection{
-				Title:  "status",
-				Type:   "status",
-				Name:   name,
-				Values: valueMap,
-			}
-			configSections = append(configSections, rt)
-		}
-	}
-
-	return configSections
-}
-
-func addConfigSections(configSections []ConfigSection, title string, sectionMap SectionMap) []ConfigSection {
-	for name, values := range sectionMap {
-		t := strings.Split(name, "/")
-		if len(t) > 2 {
-			log.Errorf("invalid section(slash), %v", t)
-			continue
-		}
-
-		values_ := values.([]map[string]interface{})
-		valueMap := make(map[string]string)
-
-		for _, m := range values_ {
-			for k, v := range m {
-				switch v.(type) {
-				case int64:
-					valueMap[k] = strconv.FormatInt(v.(int64), 10)
-				case bool:
-					valueMap[k] = strconv.FormatBool(v.(bool))
-				default:
-					valueMap[k] = v.(string)
-				}
-			}
-		}
-
-		if len(valueMap) == 0 {
-			continue
-		}
-
-		rt := ConfigSection{
-			Title:  title,
-			Type:   title,
-			Name:   t[0],
-			Values: valueMap,
-		}
-
-		if len(t) == 2 { // if args exists, store it
-			rt.Arg = t[1]
-		}
-
-		configSections = append(configSections, rt)
-	}
-
-	return configSections
-}
-
 // ValidMqttPublishTopic validates the Topic is validate or not
 // This is used with validator packages.
 func ValidMqttPublishTopic(v interface{}, param string) error {
@@ -262,69 +113,4 @@ func SearchSection(sections *[]ConfigSection, t, arg string) *ConfigSection {
 		}
 	}
 	return nil
-}
-
-// LoadConfig loads toml format file from confPath arg and returns []ConfigSection.
-// ConfigSection has a Type, Name and arg.
-// example:
-// [broker."sango"]
-// [broker."sango/1"]
-// [broker."sango/2"]
-//
-// ret = [
-//   ConfigSection{Type: "broker", Name: "sango"},
-//   ConfigSection{Type: "broker", Name: "sango", Arg: "1"},
-//   ConfigSection{Type: "broker", Name: "sango", Arg: "2"},
-// ]
-func LoadConfig(confPath string) (Config, error) {
-	dat, err := ioutil.ReadFile(confPath)
-	if err != nil {
-		return Config{}, err
-	}
-
-	return LoadConfigByte(dat)
-}
-
-// LoadConfigByte returns []ConfigSection from []byte.
-// This is invoked from LoadConfig.
-func LoadConfigByte(conf []byte) (Config, error) {
-	config := Config{}
-	var configToml ConfigToml
-
-	if err := toml.Unmarshal(conf, &configToml); err != nil {
-		return config, err
-	}
-
-	var sections []ConfigSection
-	var bn []string
-
-	// gateway section
-	config.GatewayName = getGatewayName(configToml.Gateway)
-	if config.GatewayName == "" {
-		return config, fmt.Errorf("gateway has not name")
-	}
-	sections = addGatewaySection(sections, configToml.Gateway)
-
-	// status section
-	sections = addStatusSections(sections, configToml.Status)
-
-	// broker sections
-	sections = addConfigSections(sections, "broker", configToml.Brokers)
-
-	// device sections
-	sections = addConfigSections(sections, "device", configToml.Devices)
-
-	// broker names
-	for name, _ := range configToml.Brokers {
-		t := strings.Split(name, "/")
-		if len(t) > 2 {
-			continue
-		}
-		bn = append(bn, t[0])
-	}
-
-	config.Sections = sections
-	config.BrokerNames = bn
-
-	return config, nil
 }
